Debian: Fix downloading sources
The change to support downloading submodules lost the ability to
download branches or tags. Since the script is doing that _all_
_the_ _time_ and expects certain stable points in the source
history to be used, that's a kinda crucial feature.

Reinstate that ability and re-try on a clean checkout to ensure
that it _really_ works this time.

Shout-out to Johannes Brakensiek (@Letterus) for reporting the issue.